Correct potentially bogus warning in openTiff
openTiff checks that the size of the output image correspond to the
size of the image read from the image header. It does it even if
the user specify regionSpec and doesn't load the whole image.
Do not check using the header if the image is cropped. Still check
using the regionSpec argument but I guess that there would be a crash
before if that was wrong